Modified rtc api endpoint & fixed value types

+++ components/http_server/include/http_server.h
@@ -4,5 +4,5 @@ void setup_wifi(void);
 httpd_handle_t setup_server(void);
 
 extern uint64_t watering_interval;
-extern uint16_t watering_duration;
+extern uint64_t watering_duration;
